move AbstractGenericRecord to impl package
