+++ hazelcast/src/main/java/com/hazelcast/internal/serialization/impl/compact/CompactGenericRecord.java
@@ -18,7 +18,7 @@ package com.hazelcast.internal.serialization.impl.compact;
 
 import com.hazelcast.internal.json.JsonEscape;
 import com.hazelcast.internal.serialization.impl.InternalGenericRecord;
-import com.hazelcast.nio.serialization.AbstractGenericRecord;
+import com.hazelcast.internal.serialization.impl.AbstractGenericRecord;
 
 /**
  * An extension of the {@link AbstractGenericRecord} that requires the
